+++ src/sonic_ax_impl/mibs/ietf/rfc3433.py
@@ -10,6 +10,8 @@ from ax_interface import MIBMeta, MIBUpdater, ValueType, SubtreeMIBEntry
 from sonic_ax_impl import mibs
 from sonic_ax_impl.mibs import Namespace
 
+from .physical_entity_sub_oid_generator import get_transceiver_sensor_sub_id
+
 @unique
 class EntitySensorDataType(int, Enum):
     """
@@ -318,7 +320,7 @@ class PhysicalSensorTableMIBUpdater(MIBUpdater):
                 raw_sensor_value = transceiver_dom_entry_data.get(sensor_key)
 
                 sensor = get_transceiver_sensor(sensor_key)
-                sub_id = mibs.get_transceiver_sensor_sub_id(ifindex, sensor_key)
+                sub_id = get_transceiver_sensor_sub_id(ifindex, sensor_key)
 
                 try:
                     mib_values = sensor.mib_values(raw_sensor_value)
